# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: 网络运维监控平台 API
  description: |
    网络运维监控平台后端API规范，基于RESTful设计风格。

    ## 认证方式
    - 所有API需要通过Header传递API Key：`Authorization: Bearer <token>`
    - 登录接口返回token

    ## 响应格式
    所有接口统一响应格式：
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: IT Manager Team
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: http://localhost/api
    description: 生产环境
...
paths:
...
  /auth/login:
    post:
      tags:
        - 认证
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: 用户名
                password:
                  type: string
                  description: 密码
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/logout:
    post:
      tags:
        - 认证
      summary: 用户登出
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
  /auth/skip-password-change:
    post:
      tags:
        - 认证
      summary: 跳过改密 (非首次场景)
      description: |
        C7: 用户在 `/change-password` 页可点"跳过"调本接口。
        写 `password_set_at = NOW()` + 清 `must_change_password = false` + 写 audit_log。

        主人 7/02 决策: **首次登录强改密不可跳**。
        - `reason=first_login` → 400 (seed admin/admin123 这类必须改)
        - `reason=optional` (或不传) → 允许 (用户主动取消改密, 不算"首次")
      operationId: skipPasswordChange
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  enum: [first_login, optional]
                  description: first_login=必改拒绝; optional=允许跳 (默认)
                  example: optional
      responses:
        '200':
          description: 跳过成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 首次登录必改, 拒绝跳过

  # ==================== 仪表盘 ====================
...
  /dashboard/kpis:
    get:
      tags:
        - 仪表盘
      summary: 获取关键 KPI 指标
      description: |
        返回 MTTR (平均恢复时间)、MTTD (平均检测时间)、告警密度、SLA 达成率。
        days 默认 7，范围 1-90。
      operationId: getKPIs
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 7
            minimum: 1
            maximum: 90
          description: 窗口天数，默认 7，范围 1-90
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KPI'
        '400':
          description: 请求参数错误

  # ==================== 资产管理 ====================
  /assets:
    get:
      tags:
        - 资产管理
      summary: 获取资产列表
      operationId: listAssets
      parameters:
        - name: site_id
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'

    post:
      tags:
        - 资产管理
      summary: 创建资产
      operationId: createAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
  /assets/{id}:
    get:
      tags:
        - 资产管理
      summary: 获取资产详情
      operationId: getAsset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'

    put:
      tags:
        - 资产管理
      summary: 更新资产
      operationId: updateAsset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'

    delete:
      tags:
        - 资产管理
      summary: 删除资产
      operationId: deleteAsset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功

  # ==================== 告警中心 ====================
  /alerts:
    get:
      tags:
        - 告警中心
      summary: 获取告警列表
      operationId: listAlerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [problem, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: integer
        - name: host_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertList'
  /alerts/{id}:
    get:
      tags:
        - 告警中心
      summary: 获取告警详情
      operationId: getAlert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /alerts/{id}/acknowledge:
    put:
      tags:
        - 告警中心
      summary: 确认告警
      operationId: acknowledgeAlert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 确认成功
  /alerts/{id}/resolve:
    put:
      tags:
        - 告警中心
      summary: 解决告警
      operationId: resolveAlert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 解决成功
  /alerts/{id}/mark-fp:
    post:
      tags:
        - 告警中心
      summary: 标记/反标记告警为误报
      operationId: markFalsePositive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_false_positive]
              properties:
                is_false_positive:
                  type: boolean
                  description: true=标记为误报，false=反标记
                note:
                  type: string
                  description: 误报备注（如"周期性抖动"）
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          description: 告警不存在
  /alerts/false-positives/export:
    get:
      tags:
        - 告警中心
      summary: 导出误报训练集 CSV
      operationId: exportFalsePositives
      parameters:
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: 可选 RFC3339 时间，增量导出 marked_at >= since 的记录
      responses:
        '200':
          description: CSV 文件流
          content:
            text/csv:
              schema:
                type: string
...
  /alert-rules:
    get:
      tags:
        - 告警中心
      summary: 获取告警规则列表
      operationId: listAlertRules
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRuleList'

    post:
      tags:
        - 告警中心
      summary: 创建告警规则
      operationId: createAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleInput'
      responses:
        '201':
          description: 创建成功
  /alert-rules/{id}:
    put:
      tags:
        - 告警中心
      summary: 更新告警规则
      operationId: updateAlertRule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleInput'
      responses:
        '200':
          description: 更新成功

    delete:
      tags:
        - 告警中心
      summary: 删除告警规则
      operationId: deleteAlertRule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功

  # ==================== 机房 ====================
...
  /sites/{id}:
    get:
      tags:
        - 机房机柜
      summary: 获取机房详情
      operationId: getSite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDetail'

  # ==================== 机柜 ====================
  /racks:
    get:
      tags:
        - 机房机柜
      summary: 获取机柜列表
      operationId: listRacks
      parameters:
        - name: site_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RackList'
  /racks/{id}:
    get:
      tags:
        - 机房机柜
      summary: 获取机柜详情
      operationId: getRack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rack'
  /racks/{id}/devices:
    get:
      tags:
        - 机房机柜
      summary: 获取机柜内设备
      operationId: getRackDevices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RackDeviceList'

  # ==================== 工单 ====================
  /tickets:
    get:
      tags:
        - 工单管理
      summary: 获取工单列表
      operationId: listTickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketList'

    post:
      tags:
        - 工单管理
      summary: 创建工单
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketInput'
      responses:
        '201':
          description: 创建成功
  /tickets/{id}:
    get:
      tags:
        - 工单管理
      summary: 获取工单详情
      operationId: getTicket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'

    put:
      tags:
        - 工单管理
      summary: 更新工单
      operationId: updateTicket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketInput'
      responses:
        '200':
          description: 更新成功

  # ==================== 用户 ====================
...
  /users/{id}:
    get:
      tags:
        - 用户管理
      summary: 获取用户详情
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==================== 通知渠道 ====================
  /notification/channels:
    get:
      tags:
        - 通知设置
      summary: 获取通知渠道列表
      operationId: listChannels
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelList'

    post:
      tags:
        - 通知设置
      summary: 创建通知渠道
      operationId: createChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelInput'
      responses:
        '201':
          description: 创建成功
  /notification/channels/{id}:
    put:
      tags:
        - 通知设置
      summary: 更新通知渠道
      operationId: updateChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelInput'
      responses:
        '200':
          description: 更新成功

    delete:
      tags:
        - 通知设置
      summary: 删除通知渠道
      operationId: deleteChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功
  /notification/channels/{id}/test:
    post:
      tags:
        - 通知设置
      summary: 测试通知渠道
      operationId: testChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 测试消息已发送
  /diagnostics/assets/{id}/timeline:
    get:
      tags:
        - 资产诊断
      summary: 获取资产时间线（用于故障诊断）
      description: |
        聚合 alerts / tickets / assets / asset_networks 四张表，按时间倒序返回事件流。
        含 MTTR (平均恢复时间)、open_alerts、open_tickets 等摘要。
      operationId: getAssetTimeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 30
            maximum: 365
          description: 查询窗口（天）默认 30，最大 365
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 200
            maximum: 1000
          description: 事件数上限，默认 200，最大 1000
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticTimeline'
        '400':
          description: 请求参数错误
        '404':
          description: 资产不存在
...
  /diagnostics/traceroute:
    get:
      tags:
        - 资产诊断
      summary: 资产 traceroute 网络路径
      description: |
        调用系统 traceroute 跟踪资产 IP 的网络路径。
        maxHops 默认 30，上限 64。
      operationId: tracerouteAsset
      parameters:
        - name: host
          in: query
          required: true
          schema:
            type: string
          description: 目标 host（IP 或 hostname）
        - name: maxHops
          in: query
          required: false
          schema:
            type: integer
            default: 30
            maximum: 64
          description: 最大跳数，默认 30，上限 64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TracerouteResult'
        '400':
          description: 请求参数错误
        '500':
          description: traceroute 执行失败
...
  /alert-suppressions:
    get:
      tags:
        - 告警抑制
      summary: 列出告警抑制规则
      operationId: listAlertSuppressions
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertSuppression'
    post:
      tags:
        - 告警抑制
      summary: 创建告警抑制规则
      operationId: createAlertSuppression
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertSuppression'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSuppression'
        '400':
          description: 参数错误
...
  /oncall/schedules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags:
        - 值班
      summary: 删除值班组（级联删 shifts）
      operationId: deleteOncallSchedule
      responses:
        '204':
          description: 成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string

    # ==================== 认证 ====================
...
    LoginResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          properties:
            token:
              type: string
            user:
              $ref: '#/components/schemas/User'
            # C7: 首次登录强改密 flag
            # TRUE = 需要 redirect 到 /change-password (前端检测)
            # FALSE = 正常 dashboard
            must_change_password:
              type: boolean
              example: false

    # ==================== 仪表盘 ====================
...
    AlertRuleList:
      type: object
      properties:
        code:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/AlertRule'

    # ==================== 机房机柜 ====================
...